Issue #15594:cast null as integer

+++ xtupleserver/trunk/dbscripts/functions/createpr.sql
@@ -201,7 +201,7 @@ BEGIN
   ELSIF (pParentType = 'F') THEN
     SELECT planord_itemsite_id AS itemsiteid,
            planord_qty AS qty,
-           planord_duedate AS duedate, NULL AS prjid 
+           planord_duedate AS duedate, NULL::INTEGER AS prjid 
            INTO _parent
     FROM planord
     WHERE (planord_id=pParentId);
